latogeo-back/requirements/base.txt
django==1.10.6
django-password-reset==1.0
djangorestframework==3.6.3
Markdown==2.6.8
django-filter==1.0.4
